// src/components/anfrage/types.ts
export interface AnfrageFormData {
  gebauedeTyp: string;
  dachform: string;
  dacheindeckung: string;
  dachflaeche: string;
  eigentuemer: string;
  anrede: string;
  vorname: string;
  nachname: string;
  email: string;
  telefon: string;
  plzOrt: string;
  nachricht: string;
  website_url: string;
  privacyAccepted: boolean;
}